Separation of jvm-specific code from TableSchemaSelectConnectionTest
